Move common definitions for loadfile into separate header file
